requirements.txt
numpy
scipy
pandas
matplotlib
seaborn
scikit-learn
jupyterlab
ipykernel
pyyaml
tqdm
mne
mne-bids
pyarrow
fastparquet
sqlalchemy
dash
dash-bootstrap-components
plotly
pytest
